Move time functions to the separate files.
