+++ Library/time/clock.c
@@ -1,0 +1,6 @@
+#include "../time.h"
+
+clock_t clock()
+{
+    return micros_timer_get_system_clock();
+}
