+++ Library/time/difftime.c
@@ -1,0 +1,6 @@
+#include "../time.h"
+
+double difftime(time_t end, time_t beginning)
+{
+    return end - beginning;
+}
